Updated openstack/openstack Project: openstack/cinder  21f6ce2fac004eb28dae1904c7deacf1c4131dee
VMAX driver - Manage/Unmanage performance fix

Previously managing an existing volume made a call to the ECOM
server which retrieved every volume across all connected arrays.
When this call was made if there were heavily populated arrays
the call would hang and eventually crash. To get around this,
instead of seeking the volume at the start of the manage operation,
the relevant details are filled in to a CIMInstanceName object and
this object is used for a more directed search to the ECOM server,
returning the volume successfully if it exists.

Change-Id: Ib8a628016caf88f67ec54ce10bc735053b70c8c2
Closes-Bug: #1681888