Updated openstack/openstack Project: openstack/cinder  03c85d96511c9c54630019baad115aa2cba746ba
rbd: resize volume not cloned but full copy

In case cloning was disabled by rbd_max_clone_depth=0 and only
a full copy will be created make sure the volume is resized in
case the destination volume is requested to be larger than the
source vol. Added new testcase to cover the issue.

Closes-Bug: #1559913

Co-Authored-By: Xinli Guan <xinli@us.ibm.com>

Change-Id: I4b2af6c28c04335be1bc7ec53055c43485c8cde3
Signed-off-by: Danny Al-Gaaf <danny.al-gaaf@bisect.de>