Updated openstack/openstack Project: openstack-infra/shade  ba0e9451e664743dbc5d24307ca5a3829ee3e91b
Pick most recent rather than first fixed address

If there are multiple fixed ips for a server after nat_destination
filtering, we pick one arbitrarily. The thing we really want to do is
pick the most recent one - since it's the one with the greatest chance
of success.

Removed a test that wasn't actually testing anything.

Change-Id: I73fe5fe58269931ae9a7e52b79c3211d96e69a92