Updated openstack/openstack Project: openstack/cinder  11610366396cc9a399dcf8027cfafdab7818a4d3
SMBFS: avoid unnecessary conversion when uploading volumes

When uploading vhdx images, we're always converting them to 'vhd'
first. This is an unnecessary operation that can be dropped.

This behavior was introduced along with the driver, at a time in
which qemu-img did not support vhdx images and we were opting for
vhd images for compatibility reasons.

Change-Id: I2714fcb23c3e4d61e788ef89f4dae6a72124b2ff