Updated openstack/openstack Project: openstack/nova  d0195e0699a1b0cbc16be849a4dd3aca0201387b
doc: add links to summit videos on cells v2

We've had a couple of nice presentations on cells v2
from previous summits. This change links those into the
top-level cells v2 docs in the devref. Note that I
consciously put them before the Manifesto section as
that might be somewhat dated at this point, but it would
also be nice as someone new to watch a presentation before
digging into prose for the details. I also elected to avoid
adding yet a third "References" section to this page since
the other ones are specifically references for install and
upgrade scenarios.

Change-Id: I8e60e018cd103e18dd34c3f9cbdd409ec6550a97