Updated openstack/openstack Project: openstack/nova  c19194c976664a03023d5694d29ee7daa5d7d891
doc: start a FAQs section for cells v2

We get random questions about cells v2 in IRC and some of them
are repeats, and some are just good questions which took some
thought to answer, so we should document those so they aren't
lost.

This is the start of a series, so it adds an initial question.

Change-Id: I96005d1b86bb92f71a5cc848afd2d0126dd9071a