Updated openstack/openstack Project: openstack/cinder  7b304ce4aadffcc34c80d3b15de2685fadf7056f
attach_volume first check for existing attachment

Attach_volume has logic to check for an existing
attachment and return that attachment. Moved this
logic before checking to see if the status of the
volume is 'is-use'.

If a first call to attach fails with a time out,
cinderclient will retry the operation therefore
attach_volume needs to return the already existing
attachment if it exists for this instance.

Updated testcase.

Change-Id: I86d7f01b6aca376f9d8cf9de6cdfecd436726a0e
Closes-bug: 1681543