Updated openstack/openstack Project: openstack/heat  fa27404f4e9bbd474e8965931363267b17cbde3a
Get rid of timeutils.round_to_seconds()

This was introduced in 0bc4de65e3bd9dc9902e6833373c78df0badbcb4 to try
to ensure that we never got a negative difference in times, which would
trigger a bug in our handling of time differences.

Now that bug 1688327 is fixed, there's no need for this complexity any
longer.

Change-Id: Iff76ff5ab2b1d642d464f111e78e248faf6ac77c
Related-Bug: #1688327
Related-Bug: #1493166