Updated openstack/openstack Project: openstack/cinder  2ef76fb251d75401ebad9069bc24cb632312d0aa
Remove BRCD primary_switch_wwn config option

This config option was deprecated two years ago and should be removed.

The way this was "deprecated" basically just removed the option even
though StrOpt was still there. It was not referenced anywhere. For this
reason, I don't feel a release note is warranted since it was effectively
already gone.

Change-Id: I56adc822845c04a81747969c6202290d9eb64a7b