Updated openstack/openstack Project: openstack/cinder  f64269a986f30965449bb707d4ef6a6b5d5407da
Check the volume metadata value is a string or not

When creating or updating the metadata for a volume, if user
give a nested value, cinder-api will raise a 500 error, the root
cause is a syntax error in sql statement.

According to cinder api ref [1], the metadata value should be a
string.

[1] https://developer.openstack.org/api-ref/block-storage/v3/?expanded=update-a-volume-s-metadata-detail#update-a-volume-s-metadata

Change-Id: I6febfe177fb49372bb6413d3b6478cb762170b2a
Closes-Bug: #1690363