Updated openstack/openstack Project: openstack/python-cinderclient  58c60e787f7b5a0fcc856f1114e5a2fe03beb364
Handle dashes in encryption-type-create arguments

encryption-type-create currently takes args "--key_size"
and "--control_location".

Our standard is to use dashes rather than underscores for
arguments like this, because it's easier to type them.

Additionally, encryption-type-update already uses
"--key-size" and "--control-location" for the same args.

This adds the dashed versions to the CLI and makes them the
default shown in the help.  The underscore versions are
retained for compatibility.

Also removes redundant "(Optional)" text for arguments listed
under "Optional Arguments".  This matches other commands.

Change-Id: I1bd2b7657ec577b9775eacd163cfdf6eb6b6eab2