Updated openstack/openstack Project: openstack/nova  3ce504de02eddcd8345356ec259f0a0063988b40
placement: implement get_inventory() for libvirt

Implements the get_inventory() method for the libvirt virt driver. In the
process, we add some code that merges information about allocation ratios and
reserved amounts from the compute node object into the inventory data returned
from the get_inventory() method. We're careful not to overwrite any inventory
information that was passed back from the get_inventory() call, such as is the
case with the Ironic virt driver's get_inventory() method that overrides the
allocation_ratio values to 1.0.

Change-Id: Ia74612dfb1a688a739f4537acf1eadfd29b9e8a2