Updated openstack/openstack Project: openstack/cinder  cec1d5c2493eb822a49c619d7ee62f555c3fa1ae
Add blockbox to Cinder project

This change creates a contrib directory in the Cinder tree and adds
a project called block-box.

This addition provides a method to easily create Docker images for
a standalone Cinder deployment and employs a docker-compose file to
enable rapid build and deployment of Cinder as a standlone service.

Change-Id: Id08ffefd322e80a433b58d48b8b89b1ca1695150