Updated openstack/openstack Project: openstack/python-openstackclient  4a27f51446800b31117534ce208885ea05d5457e
Use _get_token_resource in role assignment list

If project matches the project from access token,
we do not have to send an API request to /projects?name=...,
because the project ID is already known.
This API request may require additional permissions, so
we want to avoid it, if possible.

Change-Id: Ice1af8686bceea6b67229dcab7cf82eef821163e
Closes-Bug: #1658189