Updated openstack/openstack Project: openstack/nova  bedcf29e4586a755353b7844f8e11173e0892729
Add strict option to discover_hosts

This adds a --strict option that can be passed in when calling the
nova-manage cell_v2 discover_hosts command. When this option is used,
the command will only return success if a new host has been found.
In any other case it is considered a failure.

Closes-Bug: #1692982

Change-Id: I942af11a3987e1edce67423c66931ad9ece65587