Updated openstack/openstack Project: openstack/requirements  afcfdd95950aa959d11b6220567908a8eb430869
bump neutron-lib to 1.7.0

The upper constraints for neutron-lib have already been bumped for 1.7.0
in I1671b056819960fd6409a4a2fcac7b6c620a6ca4, but we'd like to see the
requirement change propagated into all consumers so that we can begin
using the changes in neutron-lib between 1.6.0 and 1.7.0. This patch
bumps neutron-lib to 1.7.0 so consuming requirements.txt will be updated
allowing us to use the latest release in neutron and other networking
projects.

Change-Id: I455ded91d3102e310bdbc8e698b066f845613044