Updated openstack/openstack Project: openstack/requirements  e09cd1817270d8a04f23f2aaee4d59a275be480c
Introducing pydot3

The oslo.automaton library supports generating svg images of a state
machine using the pydot library[1] Ironic inspector would like to use this
to generate image of its state machine for the documentation[2] Currently
ironic inspector uses custom tox env with an additional dependency[3] but
we'd like to reuse global dependencies.  The pydot3 package was selected
because pydot2 has known issues in python3[4]

Q: Is the library actively maintained?
A: Last commit from November 2016 [5]

Q: Is the library good code?
A: Single-file library, 500 sloc, no tests,
   fork of Pydot to fix python3-specific issues

Q: Is the library python 3 compatible?
A: The goal is to port pydot to python3;
   works with python3.6

Q: Is the library license compatible?
A: It is indeed; can be distributed under the MIT License conditions

Q: Is the library already packaged in the distros we target (Ubuntu latest
   / Fedora latest)?
A: No

Q: Is the function of this library already covered by other libraries in
   global-requirements.txt?
A: No, although pydotplus seems similar but oslo.automaton svg exports
   didn't work for me

Q: Is the library required for OpenStack project or related dev or
   infrastructure setup? (Answer to this should be Yes, of course) Which?
A: Yes, ironic-inspector would like to use it to be able to generate nice
   images of its state machine[2][3]

Q: Do I need to update anything else?
A: Having followed the proposing changes guidelines[6] a bump of
   click-repl===0.1.2 was observed in upper-constraints.txt

[1] https://docs.openstack.org/developer/automaton/api.html#module-automaton.converters.pydot
[2] https://github.com/openstack/ironic-inspector/blob/master/tools/states_to_dot.py
[3] https://github.com/openstack/ironic-inspector/blob/master/tox.ini#L51
[4] https://github.com/erocarrera/pydot/issues/107
[5] https://github.com/log0/pydot3
[6] https://docs.openstack.org/developer/requirements/#proposing-changes

Change-Id: I82e3651eefd3186c9e131b0db29a328967fb3161