Updated openstack/openstack Project: openstack/cinder  da08bc0f6e144f58195721a075824044815f8658
VMAX driver - Detaching volumes if part of two or more MVs

When Live migration is used extensively there can be scenarios
where a regular attached volume can belong to two or more Masking
Views. Because of this, we did not remove the volume from the storage
group, which is not typical behaviour. In this fix we use a temporary
file to determine if a terminate_connection is a regular detach or a
part of the live migration process.

Change-Id: Ide38fa21d65859a5516c577a9983124d998a2e95
Closes-Bug: #1684595