Updated openstack/openstack Project: openstack/nova  e97d6891b788540a594169e6f14074a5f159e477
Don't delete allocation if instance being scheduled

There could be a potential race if something other than the compute
service create allocations (say, the scheduler) because then the
self-healing process that runs periodically on compute would delete that
allocation given it wouldn't find the related instance for that host.

Removing an unnecessary method in the report client code, and rather do
the loop in the RT code for directly calling to delete allocations.

Co-Authored-By: Ed Leafe <ed@leafe.com>

Blueprint: placement-claims

Change-Id: I80ba844a6e0fcea89f80aa253d57ac73092773ae