Updated openstack/openstack Project: openstack/cinder  2841fbe0449d7c80e0c9c0da53a31574810376cc
SMBFS: report each share as a pool

Cinder RemoteFS based drivers report total space capacity as a
sum of all individual share capacities.

For this reason, a volume backend may be considered for placing
a volume, even if none of the shares used by the backend can actually
store the requested volume.

This patch changes this behavior for the SMBFS drivers, treating
each share as a pool and repoting the stats accordingly.

This will also allow the scheduler to target a specific share.

The pool names will be configurable. If no pool name is explicitly
mapped to a share, the share name will be used instead as a pool name.

Partially Implements: blueprint remotefs-pools-support

Change-Id: I7ad82a051ba2732f2f7e497e795aff83efcfd428