Updated openstack/openstack Project: openstack-infra/devstack-gate  d4f843c4debadf2d02626fe584a23915184a372c
Compress stackviz data

Each stackviz "install" needs about 15MB of disk, a significant portion
of which is just json data that is rendered. We can compress that as
long as our server knows to serve .json.gz when asked for .json files
(and it should, but this serves as a self testing check of that).

Change-Id: Ie0f25f06bdd539945f5f9c26885f0388354bd129