Updated openstack/openstack Project: openstack/tripleo-heat-templates  34cbc97d57ea618ec97267e24f030e02f7b124d0
Store role_data in an OS::Heat::Value resource

Looking up role_data is very slow, particularly when referencing the
RoleData output, as it re-resolves every output for all the (many) nested
stacks in the *ResourceChain resources.

There is work ongoing to optimize this in heat, but this approach improves
performance considerably (my local output-show for RoleData is 10x faster)
so we can consider including RoleData in the tripleo dynamic ansible inventory,
which may be needed for validations and minor updates in future.

Change-Id: I5e6665703e859dc1ec6b60dece70f858c9afaf66