Updated openstack/openstack Project: openstack/nova  933c2c5c5efb5618395441f3ebf1c1a8125bdd27
Avoid lazy-loading instance.id when cross_az_attach=False

The instance is no longer created in the API so the id attribute
won't be set, which means when checking the instance AZ against
the volume AZ, if they don't match we can't put the instance.id
in the error message. We shouldn't have been putting the instance
primary key in the error message anyway.

This fixes the bug by using the instance.uuid which is set in
this object in _provision_instances.

Change-Id: I396b767815b666706fec980ded482fa4746d2efc
Closes-Bug: #1693654