Updated openstack/openstack Project: openstack/nova  4ea98ab6f0f9ec5780d6fcdb794d8d7261070b4a
Optimize creating security_group

When passing a source group_id for creating a new security group rule,
the code is looking up that group twice. We should just look that up
once at the beginning and reuse it so we don't have to hit the nova db
or neutron twice for the same thing.

Change-Id: I3ea4586d11b14da5f6d93e8d20f05171e7e6f0e8
Closes-bug: #1580621