Updated openstack/openstack Project: openstack/nova  d9d300d5c80265c741f26cc3cdca3ef8d08e6e81
Migrate to oslo request_id middleware - mv 2.46

In order to support cross project request_id tracking, we need to be
on oslo.middleware for request_id. This makes that change now that
oslo middleware can support compat headers.

api-ref is not updated yet because x-compute-request-id was apparently
never documented there, and the timing on landing this is narrow
because the moment a requirements update happens we'll have this new
behavior.

Part of bp:oslo-middleware-request-id

Change-Id: I4d9f91b01de12cd0a676fc649953f98473b6b416