Updated openstack/openstack Project: openstack-infra/devstack-gate  9d61c915321c259c8c4ddf9c08050a18acc9246d
Store journal in export format not native

We do this because not all distros/release have the tool to properly
export the journal data in native format (specifically
systemd-journal-remote). We can however trivially store it in the export
format then anyone that wants to review these logs can import them to
native format locally using systemd-journal-remote.

Closes-Bug: #1691013
Change-Id: I13cdb786234e05dd3d8c581aaeb8a47d9f73b345