Updated openstack/openstack Project: openstack/cinder  cb67f4f177d232db3b8138c830c7fa2cfe0fa362
Make "connector" param to be optional in force-detach api

One of the purposes of force-detach is to help the admin clean up
if things are in a bad state, so there might not be a connector to
pass in. This is the case if Cinder thinks the volume is attached
but the Nova instance has already been deleted.

Change-Id: I417ed7b100dfe8402599028597511949e00725ac
Closes-Bug: #1619647