Updated openstack/openstack Project: openstack/nova  0d31fb303e07b7ed9f55b9c823b43e6db5153ee6
Use targeted context when burying instances in cell0

After Iccdf6b80f5fc8adcc8a89ce6ece3f37b6cbcaee2 we need to
use the yielded context which is targeted to the cell when
we do DB operations, which in this case is creating the
instance in the cell0 database and then updating it's status.

There is another place in here where this was missed, which is
when we're trying to delete a build request which was already
deleted.

Closes-Bug: #1693438

Change-Id: I142f97d691fa55e9824714c9c224f998ad72337e