Updated openstack/openstack Project: openstack/tempest  123eb2aa55b4c1bc13f4447aec5d22f9968bbe58
Fix for volume quota class test

When 'VolumeQuotaClassesTest.test_update_default_quota' is tested in race
with other tests which involve creating/deleting volume types or updating
default quotas of volume type, it may cause the assertions in
'test_update_default_quota' failed. This patch is to solve this problem.

Change-Id: I6566519ad0b19f62a3ac7365de33ed3fc62517d7
Closes-Bug: 1687488